+++ src/test/java/at/rtr/rmbt/repository/impl/RadioSignalRepositoryImplTest.java
@@ -85,7 +85,7 @@ class RadioSignalRepositoryImplTest {
         when(resultSet.getObject("signal_strength", Integer.class)).thenReturn(TestConstants.DEFAULT_SIGNAL_STRENGTH);
         when(resultSet.getString("cat_technology")).thenReturn(TestConstants.DEFAULT_CAT_TECHNOLOGY);
         when(resultSet.getObject("location_id", Long.class)).thenReturn(TestConstants.DEFAULT_LOCATION_ID);
-        when(resultSet.getObject("area_code", Integer.class)).thenReturn(TestConstants.DEFAULT_AREA_CODE);
+        when(resultSet.getObject("area_code", Long.class)).thenReturn(TestConstants.DEFAULT_AREA_CODE);
         when(resultSet.getObject("primary_scrambling_code", Integer.class)).thenReturn(TestConstants.DEFAULT_PRIMARY_SCRAMBLING_CODE);
         when(resultSet.getObject("channel_number", Integer.class)).thenReturn(TestConstants.DEFAULT_CHANNEL_NUMBER);
         when(resultSet.getObject("timing_advance", Integer.class)).thenReturn(TestConstants.DEFAULT_TIMING_ADVANCE);
